refactored lagprevention
